Adds a sender identifier to the URLs for portal scenario
The back end service needs information about for whom the message is sent on behalf of in cases where a broker sends the actual HTTP request on behalf of another sender.
This information is already in the document bundle, but the back end needs it as part of the URL to be able to check the broker/sender relationship before processing the HTTP body.