Throws different exceptions when a job can't be cancelled
If a «cancellable» doesn't have a cancellation URL it is because the job already has been completed. In these cases the client should have checked the job's status before cancelling. Completed jobs cannot be cancelled.
If the service returns 409 CONFLICT upon cancellation it is because the job has been completed *after* returning a cancellation-URL to the client but before the client has called said URL. This might happen and is not considered wrong behaviour from the client.