Senders may specify for how many seconds the document is available to sign rather than a single point in time where it expires
This has been changed to a relative value to be able to support "chained signatures" (i.e. one signer is required to sign before another) in the future. In such a case it doesn't make sense to specify a single point in time where the document expires, so the sender may specify a duration the document is available _for each signer_.