Allows specifying sender per job
This is useful in cases where you're acting as a broker on behalf of multiple organisations and needs a way to set different senders on different jobs.
Will fall back to use the sender optionally specified during instantiation of the client if no sender is specified for a job to support cases where one don't need to act on behalf of multiple senders.
Fixes #23