Clean up construction of Portal and Direct jobs
Reduce number of overloaded builder constructors: keep mainly one
constructor with lists for both documents and signers, and only one
additional for convenience for exactly one document and one signer.
Maintaining overloaded methods with vararg variants for multiple list parameters
does not scale, and it is easy to just wrap a single element in either
Arrays.asList (Java 8) or List.of (Java >= 9).