Trust Buypass SEID2 enterprise certificates
Modifies the alias-generating for certificate to include the name of the containing
folder, because the file names for test and production CA certs are the
same, and to be able to reliably test that the trusts are containing the
expected certifiates. The aliases are not used by the client.

Add test that we recognize orgnr in SEID2 certs