Extract http client config to separate classes
Introduces and API for configuring timeouts, TimeoutsConfigurer, and
a separate API, ApacheHttpClientConfigurer, for accessing the configuration facilities of the
underlying Apache HTTP Client. The last one may be used as a last resort
for fine tweaking the http client, and the library does not offer the
required configuration API.

The "global sender" concept has been renamed to "default sender".