Configurable connection pool size
Condensing the configured size to mean both total max connections and connections per
route, as the client is only concerned about one route, i.e. the target
host of the API (including any intermediate proxies).

Setting the default to be 10. The previous (fixed) default was 25 total
connections, but a maximum of 5 was allocated for the route.